+++ application/components/SIG/SIG_Hub.h
@@ -16,8 +16,8 @@ public:
     inline SIG_Hub()
     {}
 
-    void toFld(size_t id, INT32 state, INT32 speed) const;
-    void toGui(size_t id, INT32 state, INT32 speed) const;
+    void toFld(size_t id, UINT8 state, UINT8 speed) const;
+    void toGui(size_t id, UINT8 state, UINT8 speed) const;
 
     void fromDsp(size_t pos, const FldState& tele) const;
     void fromDsp(size_t pos, const GuiCmd&   tele) const;
