+++ application/components/SIG/src/SIG_Hub.cpp
@@ -27,13 +27,13 @@ void SIG_Hub::fromDsp(const size_t pos, const GuiCmd&   tele) const
     {pass();}
 }
 
-void SIG_Hub::toFld(const size_t id, const INT32 state, const INT32 speed) const
+void SIG_Hub::toFld(const size_t id, const UINT8 state, const UINT8 speed) const
 {
     // CmdFld tele(state, speed);
     IL::getDispatcher().dispatch(id, CmdFld(state, speed));
 }
 
-void SIG_Hub::toGui(const size_t id, const INT32 state, const INT32 speed) const
+void SIG_Hub::toGui(const size_t id, const UINT8 state, const UINT8 speed) const
 {
     // StateGui tele(state, speed);
     IL::getDispatcher().dispatch(id, StateGui(state, speed));
