+++ application/components/TSW/src/TSW_Hub.cpp
@@ -27,12 +27,12 @@ void TSW_Hub::fromDsp(const size_t pos, const GuiCmd& tele) const
     {pass();}
 }
 
-void TSW_Hub::toFld(const size_t id, const INT32 cmd) const
+void TSW_Hub::toFld(const size_t id, const UINT8 cmd) const
 {
     IL::getDispatcher().dispatch(id, CmdFld(cmd));
 }
 
-void TSW_Hub::toGui(const size_t id, const INT32 state) const
+void TSW_Hub::toGui(const size_t id, const UINT8 state) const
 {
     IL::getDispatcher().dispatch(id, StateGui(state));
 }
