+++ specification/ifs/I_SIG_Hub.h
@@ -11,8 +11,8 @@
 class I_SIG_Hub
 {
 public:
-    virtual void toFld(size_t id, INT32 state, INT32 speed) const = 0;
-    virtual void toGui(size_t id, INT32 state, INT32 speed) const = 0;
+    virtual void toFld(size_t id, UINT8 state, UINT8 speed) const = 0;
+    virtual void toGui(size_t id, UINT8 state, UINT8 speed) const = 0;
 
     virtual void fromDsp(size_t pos, const FldState& tele) const = 0;
     virtual void fromDsp(size_t pos, const GuiCmd&   tele) const = 0;
