+++ testing/testenv/mocks/M_SIG.h
@@ -17,25 +17,25 @@ namespace test
     public:
         MOCK_CON(SIG)
         
-        inline void fromFld(INT32 state, INT32 speed)
+        inline void fromFld(UINT8 state, UINT8 speed)
         {
             call("fromFld").PARAM(state).PARAM(speed);
         }
-        inline void expectFromFld(INT32 state, INT32 speed = 0)
+        inline void expectFromFld(UINT8 state, UINT8 speed = 0)
         {
             expect("fromFld").PARAM(state).PARAM(speed);
         }
 
-        inline void fromGui(INT32 state, INT32 speed)
+        inline void fromGui(UINT8 state, UINT8 speed)
         {
             call("fromGui").PARAM(state).PARAM(speed);
         }
-        inline void expectFromGui(INT32 state, INT32 speed = 0)
+        inline void expectFromGui(UINT8 state, UINT8 speed = 0)
         {
             expect("fromGui").PARAM(state).PARAM(speed);
         }
 
-        inline INT32 type() const
+        inline UINT8 type() const
         {
             return SIG_TYPE_H;
         }
